+++ backend/src/common/decorators/auth/jwt-auth.decorator.ts
@@ -1,0 +1,20 @@
+import { JwtAuthGuard } from '@auth/infrastructure/guards/jwt-auth.guard';
+import { applyDecorators, UseGuards } from '@nestjs/common';
+import { ApiBearerAuth } from '@nestjs/swagger';
+
+/**
+ * Декоратор для защиты роутов JWT аутентификацией
+ * Комбинирует @UseGuards(JwtAuthGuard) и @ApiBearerAuth() для Swagger
+ *
+ * @example
+ * ```typescript
+ * @Get()
+ * @JwtAuth()
+ * async findAll() {
+ *   return this.service.findAll();
+ * }
+ * ```
+ */
+export const JwtAuth = () => {
+    return applyDecorators(UseGuards(JwtAuthGuard), ApiBearerAuth());
+};